back/requirements.txt
amqp==2.6.1
anyio==3.2.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.0.7
cx-Oracle==8.2.0
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.2.1
django-debug-toolbar==3.1.1
django-decouple==2.1
django-discover-runner==1.0
django-filter==2.2.0
django-localflavor==3.0.1
django-recaptcha==2.0.6
django-role-permissions==3.1.1
django-test-without-migrations==0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.2
filelock==3.3.0
gunicorn==20.0.4
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.9
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.3.1
platformdirs==2.4.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==2.3.0
pylint==2.4.4
pysendpulse==0.1.2
python-decouple==3.4
python-slugify==5.0.2
python-status==1.0.1
python-stdnum==1.13
python3-memcached==1.51
pytz==2019.3
rcssmin==1.0.6
redis==3.3.8
requests==2.23.0
rfc3986==1.5.0
rjsmin==1.1.0
sentry-sdk==0.14.4
simple-rest-client==1.0.5
simplejson==3.17.2
six==1.14.0
sniffio==1.2.0
sqlparse==0.3.1
static3==0.7.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
Unipath==1.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.8.1
websockets==9.1
whitenoise==5.3.0
wrapt==1.11.2
